+++ tinygrad/__init__.py
@@ -1,4 +1,8 @@
 from tinygrad.tensor import Tensor            # noqa: F401
 from tinygrad.jit import TinyJit              # noqa: F401
 from tinygrad.shape.symbolic import Variable  # noqa: F401
+from tinygrad.helpers import dtypes           # noqa: F401
+
+# NOTE: these should not be relied on to be stable
+from tinygrad.ops import Device               # noqa: F401
 from tinygrad.helpers import GlobalCounters   # noqa: F401
